Add files via upload
We will conduct our own data analysis and create a file to share that documents our findings. we will start by taking a look at the dataset from Udacity and brainstorming what questions we could answer using it. Then we will use pandas and NumPy to answer the questions we are most interested in, and create a report sharing the answers. we will not be required to use inferential statistics or machine learning to complete this project, but we will make it clear in our communications that our findings are tentative. This project is open-ended in that we are not looking for one right answer.